Fix conflicts and refactor code
