--- Assets/Scripts/Assemblydef.asmdef
@@ -1,3 +1,0 @@
-﻿{
-	"name": "NewAssembly"
-}
