--- Assets/Scripts/Assemblydef.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5777cb1aad1625d48a8d26d54066a504
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
