--- Assets/Tests/Tests.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4e98f811c45b65c41b35eb3bc3d2a9f2
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
